Extract the rendered entity ID codec out of renderedEntities
`renderedEntities.ts` had accumulated four unrelated concerns: a pure string ID
codec, the canvas visibility atom, the style scope, and the Cytoscape render
hooks. Only the last is what the module name describes.

The codec has no coupling to the rest — no React, no Jotai, no state — and is
exactly the unit the first four test blocks covered, so it moves out whole along
with them. 263 lines down to 193.
